Change float to double

+++ exercises/matthew_cumber/ex02_oo_basics/src/Circle.cpp
@@ -6,7 +6,7 @@
 /** Creates an instance of a Circle 
  * @param _radius The value of the radius of the circle
  */
-Circle::Circle(float _radius)
+Circle::Circle(double _radius)
 {
   radius = _radius;
   type = "Circle";
@@ -16,7 +16,7 @@ Circle::Circle(float _radius)
 /** Calculates the area of a circle
  * @return Value of area
  */
-float Circle::area()
+double Circle::area()
 {
   return M_PI*pow(radius,2);
 }
@@ -24,7 +24,7 @@ float Circle::area()
 /** Calculates the perimeter of a circle
  * @return Value of perimeter
  */
-float Circle::perimeter()
+double Circle::perimeter()
 {
   return M_PI*2.0*radius;
 }
